refactor, split some tests into separate files
