+++ src/Microsoft.DotNet.Interactive.Tests/KernelRoutingTests.cs
@@ -4,11 +4,11 @@
 using System;
 using System.Collections.Generic;
 using System.CommandLine;
-using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
 using FluentAssertions;
-using Microsoft.DotNet.Interactive.App;
 using Microsoft.DotNet.Interactive.Commands;
+using Microsoft.DotNet.Interactive.CSharp;
 using Microsoft.DotNet.Interactive.Events;
 using Microsoft.DotNet.Interactive.Tests.Utility;
 using Pocket;
